Improve ReadMe and Warning information for `vendor` upstream linking (#131)
* README: Mention version attribute unsupported in bender vendor

* Fix error when using bender vendor with version reference

* vendor: Raise warning on file missing upstream

* Update Changelog

---------

+++ src/cmd/vendor.rs
@@ -88,12 +88,8 @@ pub fn run(sess: &Session, matches: &ArgMatches) -> Result<()> {
             DependencySource::Git(ref url) => {
                 let git = Git::new(tmp_path, &sess.config.git);
                 rt.block_on(async {
-                    // TODO MICHAERO: May need throttle
-                    future::lazy(|_| {
-                        stageln!("Cloning", "{} ({})", vendor_package.name, url);
-                        Ok(())
-                    })
-                    .and_then(|_| git.spawn_with(|c| c.arg("clone").arg(url).arg(".")))
+                    stageln!("Cloning", "{} ({})", vendor_package.name, url);
+                    git.spawn_with(|c| c.arg("clone").arg(url).arg("."))
                     .map_err(move |cause| {
                         if url.contains("git@") {
                             warnln!("Please ensure your public ssh key is added to the git server.");
@@ -103,24 +99,17 @@ pub fn run(sess: &Session, matches: &ArgMatches) -> Result<()> {
                             format!("Failed to initialize git database in {:?}.", tmp_path),
                             cause,
                         )
-                    })
-                    .and_then(|_| git.spawn_with(|c| c.arg("checkout").arg(match vendor_package.upstream {
-                        config::Dependency::GitRevision(_, ref rev) => rev,
-                        // config::Dependency::GitVersion(_, ref ver) => ver.to_str(),
-                        _ => unimplemented!(),
-                    })))
-                    .and_then(|_| async {
-                        let rev_hash = match vendor_package.upstream {
-                            config::Dependency::GitRevision(_, ref rev) => rev,
-                            _ => unimplemented!(),
-                        };
-                        if *rev_hash != git.spawn_with(|c| c.arg("rev-parse").arg("--verify").arg(format!("{}^{{commit}}", rev_hash))).await?.trim_end_matches('\n') {
-                            Err(Error::new("Please ensure your vendor reference is a commit hash to avoid upstream changes impacting your checkout"))
-                        } else {
-                            Ok(())
-                        }
-                    })
-                    .await
+                    }).await?;
+                    let rev_hash = match vendor_package.upstream {
+                        config::Dependency::GitRevision(_, ref rev) => Ok(rev),
+                        _ => Err(Error::new("Please ensure your vendor reference is a commit hash to avoid upstream changes impacting your checkout")),
+                    }?;
+                    git.spawn_with(|c| c.arg("checkout").arg(rev_hash)).await?;
+                    if *rev_hash != git.spawn_with(|c| c.arg("rev-parse").arg("--verify").arg(format!("{}^{{commit}}", rev_hash))).await?.trim_end_matches('\n') {
+                        Err(Error::new("Please ensure your vendor reference is a commit hash to avoid upstream changes impacting your checkout"))
+                    } else {
+                        Ok(())
+                    }
                 })?;
 
                 tmp_path.to_path_buf()
@@ -290,8 +279,15 @@ pub fn init(
         apply_patches(rt, git, vendor_package.name.clone(), patch_link.clone())?;
     }
 
+    // Check if includes exist
+    for path in vendor_package.include_from_upstream.clone() {
+        if !PathBuf::from(extend_paths(&[path.clone()], dep_path)[0].clone()).exists() {
+            warnln!("{} not found in upstream, continuing.", path);
+        }
+    }
+
     // Copy src to dst recursively.
-    match &patch_link.from_prefix.prefix_paths(dep_path).is_dir() {
+    match link_from.is_dir() {
         true => copy_recursively(
             &link_from,
             &link_to,
@@ -304,16 +300,23 @@ pub fn init(
                 .collect(),
         )?,
         false => {
-            std::fs::copy(&link_from, &link_to).map_err(|cause| {
-                Error::chain(
-                    format!(
-                        "Failed to copy {} to {}.",
-                        link_from.to_str().unwrap(),
-                        link_to.to_str().unwrap(),
-                    ),
-                    cause,
-                )
-            })?;
+            if link_from.exists() {
+                std::fs::copy(&link_from, &link_to).map_err(|cause| {
+                    Error::chain(
+                        format!(
+                            "Failed to copy {} to {}.",
+                            link_from.to_str().unwrap(),
+                            link_to.to_str().unwrap(),
+                        ),
+                        cause,
+                    )
+                })?;
+            } else {
+                warnln!(
+                    "{} not found in upstream, continuing.",
+                    link_from.to_str().unwrap()
+                );
+            }
         }
     };
